Tie the lifetime of env rebinding records to their analysis context's
For all computations related to a given analysis, env rebinding records
are now allocated once, and deallocated only when the analysis context
itself is deallocated. These records are recycled when the rebinding
logically dies.

This is the first step towards adding safety checks for rebindings in
the public Ada API: once the analysis context of a node has been checked
as valid, we can now assume that the node's rebindings pointer is valid.

TN: T121-033